Sizable Changes
I did some more things with the CSS and and added a gradient to the submission part of the ballot (Its not perfect so if someone want to work on it be my guest. I also moved the delete buttons to the choices themselves, so that we didnt have the Xs out in the open. It works well and doesnt cause errors (atleast to my knowledge) What I would want is to get rid of that top white bar if possible, im not too used to using grid.